Set Up Initial Enemy and Item Spawn
Reworked Room Data script to include a list of enemies. These are added to the prefabs and spawn in with the room change. Will rework to only enable them when player is in the room in later versions.

Reworked Game manager to pull the enemies from the current room. Game manager keeps track of which enemies are a live and locks doors when enemy count is > 0.

Player is able to destroy enemy classes when attacking them. Good for initial weapon type, may change this to ranged weapons down the road as melee distance is a bit nippy.

Added Item Spawner script. Handles dropped items as well as items for Shop rooms. Each will contain a pool of items which could spawn. Plan for higher chance of upgrades from Shop than in drops.
Integrated into Game manager for when the player kills an enemy or clears a room. Chance of item drops for both conditions.

Misc updates:
Fixed issue with attack animation playing twice between attacks.