Updated Room Generator and Item spawns for Shops
Updated the Room generation script to place the boss room the furthest away from the player spawn. This is done first with the reference of the room being removed from the generator to prevent overwriting by a shop.

Shop items now have a cost displayed over them. If the player can't afford the item they walk into it like they would a wall.

Fixed a bug with the final boss room preventing boss enemies from detecting the player. On being enabled, the player is attached to the script.

Game resets with a new dungeon upon clearing the level or through the player dying.